periodic sync added for #22
