+++ web/templates/baseTemplate.html
@@ -46,7 +46,6 @@
                 <li>
                   <a class="dropdown-item" data-toggle="tooltip" data-placement="bottom" title="Logout" href="{{ url_for('auth.logout') }}"><i class="bi bi-person" width="16" heigth="16"></i>Logout</a>
                 </li>
-                <li><a class="dropdown-item" href="#">Item</a></li>
                 <li>
                     <hr class="dropdown-divider">
                 </li>
